requirements.txt
openai>=1.0.0
graphviz>=0.20.1
pyyaml>=6.0
tiktoken>=0.3.3
pytest>=7.0.0
python-dotenv>=1.0.0 